+++ src/Nacara.Layout.Standard/Source/Page.WithMenuOrToc.fs
@@ -278,7 +278,7 @@ let private renderTableOfContentOnly
         ]
     ]
 
-let rec private tryFindTitlePathToCurrentPage
+let rec tryFindTitlePathToCurrentPage
     (pageContext : PageContext)
     (acc : string list)
     (menu : Menu) =
@@ -310,7 +310,7 @@ let rec private tryFindTitlePathToCurrentPage
     | [ ] ->
         None
 
-let private renderBreadcrumbItems (items : string list) =
+let renderBreadcrumbItems (items : string list) =
     items
     |> List.map (fun item ->
         Html.li [