+++ src/Nacara.Layout.Standard/CHANGELOG.md
@@ -6,6 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+* PR #87: Add breadcrumb on desktop (by @mabasic)
+* PR #87: Add current section name in the navbar on mobile too
+
+### Changed
+
+* PR #86: Make the Edit button a bit more discrete (by @mabasic)
+* PR #85: Change the Navigation buttons, to be styled as "Fat buttons" now (by @mabasic)
+
 ## 1.0.0-beta-009 - 2021-08-23
 
 ### Changed